requirements.txt
fastmcp>=2.8.0
google-adk>=1.2.1
flask>=2.3.0
python-dotenv>=1.0.0
requests>=2.31.0
asyncio-mqtt>=0.16.0
tenacity>=8.2.0
uvicorn>=0.24.0
httpx>=0.25.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
